requirements.txt
fastapi
uvicorn
requests
pusher
selenium
python-dotenv
